crimes-api/pyproject.toml
[tool.poetry]
name = "crimes-api"
version = "0.1.0"
description = ""
authors = ["harsh8398 <pharsh58@gmail.com>", "ShreyaTalati <talatishreya2@gmail.com"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
fastapi = "^0.71.0"
uvicorn = {extras = ["standard"], version = "^0.17.0"}
google-auth = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
